FIX: PrologAtom unification bug fixed.
Before: The two Atoms names' single quotes were removed (if any) and the results were compared. If they are the same a new Atom is created with the new name. The problem is that, if the names were supposed to have single quotes, they loose them and the new name is invalid.

Now: After comparing the new names, additional check is made to see if the name is a single word of english letters and underlines. If it's not, the single quotes are added back to it so it becomes a valid Atom name.