>> NINE DAYS LATER <<
Due to the lack of Internet connection, no change was commited - many things changed though. All Terms will now use multimethods instead of a protocol, the Variables unification method is a bit modified and the Interpreter is working more like a native language - the terms are proved with recursion instead of changing the initial query.