ADD: Interpreter replace multimethod.
The replace method is used to replace the current Term from the query with the one, returned as a result of the resolvation of the Term and a clause from the Knowledge Base. This replacement is used to create a 'current' version of the query. With that version can be created a Checkpoint.