FIX: Some bugs fixed:
Whitespace removal bug: When removing whitespace, the parser must skip those left between two numbers or two letters: "X Y", "2 4", "a 4", because there is a missing operator there. Otherwise the two substrings will be conjoined and no error may appear.

Parsing whitespaces: When the parser reaches a whitespace, that is a whitespace left beheind for a purpouse - a missing operator. The parser throws an exception that shows the place of the missing operator.