ADD: Operators clashing.
When two operators have different precedence, the parser handles them
according to the Shunting-Yard algorithm. When they have the same
precedence, they are compared by their type (xfx, yx, etc.). If the two
types could be parsed together, they are handled with the Shunting-Yard
algorithm, but when they cannot - the Operators "clash" with each other:
That means that the parssing cannot proceed and the two operators must be
separated with brackets in order to work.