NEW: Faster PrologVariables binding.
The new logic is based on the disjoint set data structure. When two
PrologVariables are bound to one another, the right becomes a child of
the left. If we are binding the roots of large bounds-trees, the right
root becomes a child of the left. If we are binding variables,
that are not the roots of their respecitive bounds-trees, the roots are
found recursively and they are bound instead.

This trees are built very fast, but the iterations inside them are slow,
so every time we seek the root of a tree, all variables that are in the
chain of variables to the root (moving from the leaves up), are moved up so
that they become direct children of the root variable. If we iterate
more often, the tree will become more and more short and the iterations will
become with constant complexity.

Only the root can point to a value. If the root points to a value, all variables from the same bounds-tree point to the same value, given
that they take it from the root variable.