ADD: PrologFact.
PrologFact is a pair of PrologAtom and PrologArgsList. Two PrologFacts
unify if their atoms and their arguments unify.

Create method added. It creates IPrologTerms from Clojure objects (e.g.
a PrologVariable from "X"). It recognizes the type of object to be
created and creates it.

PrologFacts are created by calling (create [:fact <atom_name>
[<args>]]). The function automaticaly creates the PrologAtom and each of
the terms inside the arguments list (as well as the list itself).