Adding query solver
The solver executes queries based on the data of a given database. The query is
in the form of a Term. The solver checks if that term unifies with other terms
from the given database and produces a lazy sequence of environments that are
the results of all unifications that succeeded.